complete the project crud

+++ src/main/java/org/example/javafxbaladity/gui/ProjectController.java
@@ -21,9 +21,10 @@ import javafx.stage.StageStyle;
 import org.example.javafxbaladity.Main;
 import org.example.javafxbaladity.models.Project;
 import org.example.javafxbaladity.services.ProjectService;
+import org.example.javafxbaladity.utils.Modals;
 
 import java.io.IOException;
-import java.time.LocalDateTime;
+import java.sql.Date;
 
 public class ProjectController {
 
@@ -31,44 +32,31 @@ public class ProjectController {
     Project selectedProject = null;
 
     @FXML
+    private TableView<Project> Projects_Table;
+
+    @FXML
     private TableColumn<Project, Float> budget_col;
 
     @FXML
-    private TableColumn<Project, LocalDateTime> date_debut_col;
+    private TableColumn<Project, Date> date_debut_col, date_fin_col;
 
     @FXML
-    private TableColumn<Project, LocalDateTime> date_fin_col;
-
-    @FXML
-    private TableColumn<Project, String> description_col;
+    private TableColumn<Project, String> title_col, description_col, statut_col;
 
     @FXML
     private TableColumn<Project, Integer> id_col;
 
     @FXML
-    private TableColumn<Project, String> statut_col;
+    private Text numOfProjects, numOfActiveProjects, numOfInactiveProjects, statut_txt;
 
     @FXML
-    private TableColumn<Project, String> title_col;
+    private Button Home_Btn, User_Btn, Document_Btn, Project_Button, Reclamation_Btn, Association_Btn, Event_Btn, detailsProject_Btn, deleteProject_Btn, updateProject_Btn, open_modal_btn, statut_btn;
 
     @FXML
-    private Text numOfProjects;
-
-    @FXML
-    private Text numOfActiveProjects;
-
-    @FXML
-    private Text numOfInactiveProjects;
-
-    @FXML
-    private Button Home_Btn, User_Btn, Document_Btn, Project_Button, Reclamation_Btn, Association_Btn, Event_Btn, detailsProject_Btn, deleteProject_Btn, updateProject_Btn, open_modal_btn;
-
-    @FXML
-    private TableView<Project> Projects_Table;
-
-    @FXML
-    void onHomeButtonClick(ActionEvent event) {
-
+    void onHomeButtonClick(ActionEvent event) throws Exception {
+        FXMLLoader loader = new FXMLLoader(Main.class.getResource("views/home.fxml"));
+        Stage stage = (Stage) Home_Btn.getScene().getWindow();
+        stage.setScene(new Scene(loader.load()));
     }
 
     @FXML
@@ -97,11 +85,7 @@ public class ProjectController {
     }
 
     @FXML
-    void onUpdateProject(ActionEvent event) {
-    }
-
-    @FXML
-    void onDeleteProject(ActionEvent event) throws IOException {
+    void onUpdateProject(ActionEvent event) throws Exception {
         if (selectedProject == null) {
             Stage stage = new Stage();
             stage.setTitle("Error");
@@ -115,6 +99,36 @@ public class ProjectController {
             stage.setY(400);
             return;
         }
+
+        BoxBlur blur = new BoxBlur(3, 3, 3);
+        Node source = (Node) event.getSource();
+        Stage primaryStage = (Stage) source.getScene().getWindow();
+
+        primaryStage.getScene().getRoot().setEffect(blur);
+
+        FXMLLoader loader = new FXMLLoader(Main.class.getResource("views/project/editProject.fxml"));
+
+        Stage stage = new Stage();
+        stage.initModality(Modality.APPLICATION_MODAL);
+        stage.initStyle(StageStyle.TRANSPARENT);
+        stage.setResizable(false);
+        stage.setScene(new Scene(loader.load()));
+
+        EditProjectController editProjectController = loader.getController();
+        editProjectController.setSelectedProject(selectedProject);
+
+        stage.showAndWait();
+        showProjects();
+
+        primaryStage.getScene().getRoot().setEffect(null);
+    }
+
+    @FXML
+    void onDeleteProject() throws IOException, Exception {
+        if (selectedProject == null) {
+            Modals.displayError("Please select a project to delete");
+            return;
+        }
         projectService.delete(selectedProject.getId());
         showProjects();
     }
@@ -125,7 +139,7 @@ public class ProjectController {
     }
 
     @FXML
-    void onOpenAddModal(ActionEvent event) throws IOException {
+    void onOpenAddModal(ActionEvent event) throws Exception {
         BoxBlur blur = new BoxBlur(3, 3, 3);
         Node source = (Node) event.getSource();
         Stage primaryStage = (Stage) source.getScene().getWindow();
@@ -142,16 +156,19 @@ public class ProjectController {
         stage.setScene(new Scene(root));
 
         stage.showAndWait();
-
+        showProjects();
         primaryStage.getScene().getRoot().setEffect(null);
     }
 
     @FXML
     void getProjectRow(MouseEvent event) {
         selectedProject = Projects_Table.getSelectionModel().getSelectedItem();
+        statut_txt.setText("Projet (" + selectedProject.getTitre() +"): " + selectedProject.getStatut());
+        statut_btn.setDisable(false);
+        statut_btn.setText(selectedProject.getStatut().equals("active") ? "Terminer" : "Activer");
     }
 
-    public void showProjects() {
+    public void showProjects() throws Exception {
         ObservableList<Project> projects = FXCollections.observableArrayList(projectService.readAll());
 
         Projects_Table.setItems(projects);
@@ -162,14 +179,36 @@ public class ProjectController {
         date_fin_col.setCellValueFactory(new PropertyValueFactory<>("date_fin"));
         budget_col.setCellValueFactory(new PropertyValueFactory<>("budget"));
         statut_col.setCellValueFactory(new PropertyValueFactory<>("statut"));
-
     }
 
-    public void initialize() {
+    @FXML
+    void changeStatut(ActionEvent event) throws Exception {
+        if (selectedProject == null) {
+            Modals.displayError("Please select a project to change its status");
+            return;
+        }
+
+        if (selectedProject.getStatut().equals("active")) {
+            projectService.updateStatut(selectedProject.getId(), "inactive");
+        } else {
+            projectService.updateStatut(selectedProject.getId(), "active");
+        }
+
+        showProjects();
+
+        selectedProject = null;
+        statut_txt.setText("");
+        statut_btn.setText("Statut");
+        statut_btn.setDisable(true);
+    }
+
+
+    public void initialize() throws Exception {
         showProjects();
         numOfProjects.setText(String.valueOf(projectService.readAll().size()));
         numOfActiveProjects.setText(String.valueOf(projectService.readActiveProjects().size()));
         numOfInactiveProjects.setText(String.valueOf(projectService.readInactiveProjects().size()));
+        statut_btn.setDisable(true);
     }
